Prevent leak of magic keys in trial config
