+++ python/ray/tune/config_parser.py
@@ -190,6 +190,7 @@ def create_trial_from_spec(spec, output_path, parser, **trial_kwargs):
         restore_path=spec.get("restore"),
         trial_name_creator=spec.get("trial_name_creator"),
         loggers=spec.get("loggers"),
+        log_to_file=spec.get("log_to_file"),
         # str(None) doesn't create None
         sync_to_driver_fn=spec.get("sync_to_driver"),
         max_failures=args.max_failures,
